requirements.txt
amqp==5.2.0
asgiref==3.7.2
Babel==2.14.0
beautifulsoup4==4.12.2
billiard==4.2.0
celery==5.3.6
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap4==2023.1
crispy-bootstrap5==2023.10
Django==5.0
django-crispy-forms==2.1
django-extensions==3.2.3
django-filter==23.5
django-money==3.4.1
django-phone-field==1.8.1
django-simple-history==3.4.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
docopt==0.6.2
et-xmlfile==1.1.0
gunicorn==21.2.0
kombu==5.3.4
mysql-connector-python==8.2.0
mysqlclient==2.2.1
num2words==0.5.13
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
plotly==5.18.0
prompt-toolkit==3.0.43
protobuf==4.21.12
py-moneyed==3.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
qrcode==7.4.2
reportlab==4.0.8
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tenacity==8.2.3
typing_extensions==4.9.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.12
